requirements.txt
PySide6>=6.7.0
psutil>=5.9.0
Pillow>=10.0.0
requests>=2.31.0
rembg>=2.0.61
pillow-avif-plugin>=1.5.0
